+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Full documentation for rocFFT is available at [rocfft.readthedocs.io](https://ro
 ## [(Unreleased) rocFFT 1.0.9 for ROCm 4.0.0]
 
 ### Changed
+
+- rocFFT now automatically allocates a work buffer if the plan
+  requires one but none is provided
+
 - An explicit `rocfft_status_invalid_work_buffer` error is now
-  returned when a sufficient work buffer is required but not
-  provided.
+  returned when a work buffer of insufficient size is
+  explicitly provided.